Imported classes I need. Built the generic NN and support classes.
Currently building the Problem class that creates the input and output layers to train the NN.

VERY data intense. Going to probably need access to nearly ALL Snapshots in the database, multiple gigs of data. Need to experiment with caching or retrieving the data more efficiently. Pretty sure I am making many duplicate requests to the db which is costing too much time.