Finished save/load. Saves files with names specific to the relevant market and contract id's
Trained the NN specific to this contract and am getting very low average costs, which is good hopefully.

TODO: Build an evaluator that can take the latest market data and predict a buyYes and buyNo from it. Repeat until we have many predictions as well as recorded what the actual value ended up being.