Finished creating context menu for Mover scene's tableview. Built ContractHistory scene and am working on controller/service classes to manage that scene's functionality and data displayed.
Other scenes can open the ContractHistory scene in a new window (stage) and it creates a proper title. Needs to get the contract history data object from the service class and build the chart+tableview using that data. Have some research to do on chart building but other stuff should be fine.